requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6                # via flask
Flask==0.11.1
itsdangerous==0.24        # via flask
Jinja2==2.8               # via flask
MarkupSafe==0.23          # via jinja2
pylast==1.6.0
python-mpd==0.3.0
six==1.10.0               # via pylast
Werkzeug==0.11.11         # via flask
